requirements.txt
flask==3.0.3
python-telegram-bot==13.15
requests==2.32.3